Parameter & others: Add parameter distribution. Now it's possible to set per-voice distribution of parameter, currently only linear.
Except this, added Freq. Shift into Generator, fixed WT going out of bounds, and etc.

+++ processor/parameters/parameter_table.hpp
@@ -70,7 +70,7 @@ public:
 				return i->set_value(val);
 	}
 
-	float get_raw_value(const std::string& name)
+	float8 get_raw_value(const std::string& name)
 	{
 		for (auto& i : parameters)
 			if (i->get_name() == name)
